feat(PSG-4537): change static objects to classes, init w app id
